feature: Refactored the dates of medicos
