+++ src/main/java/com/dafon/medpluz/config/AppConfig.java
@@ -1,8 +1,7 @@
 package com.dafon.medpluz.config;
 
 import io.github.cdimascio.dotenv.Dotenv;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.*;
 
 @Configuration
 public class AppConfig {
