+++ src/main/java/com/dafon/medpluz/enums/Especialidade.java
@@ -1,4 +1,4 @@
-package com.dafon.medpluz.controller.enums;
+package com.dafon.medpluz.enums;
 
 public enum Especialidade {
     ORTOPEDIA,
