+++ src/main/java/com/dafon/medpluz/repository/MedicoRepository.java
@@ -1,7 +1,11 @@
 package com.dafon.medpluz.repository;
 
 import com.dafon.medpluz.model.Medico;
+import org.springframework.data.domain.*;
 import org.springframework.data.jpa.repository.JpaRepository;
 
+import java.util.Optional;
+
 public interface MedicoRepository extends JpaRepository<Medico, Long> {
+    Page<Medico> findAllByAtivoTrue(Pageable paginacao);
 }
