+++ src/main/resources/db/migration/V3__add-field-medicos.sql
@@ -1,0 +1,2 @@
+ALTER TABLE medicos ADD COLUMN ativo boolean;
+UPDATE medicos SET ativo = true;
